requirements.txt
h5py===2.10.0
ffmpeg_python===0.2.0
opencv_python===4.2.0.34
tensorboardX===2.0
coloredlogs===14.0
gspread===3.6.0
oauth2client===4.1.3
numpy===1.18.4
torch===1.5.1
torchvision===0.6.1
ipdb==0.13.4

